working unit tests, more protocol

+++ src/main/scala/org/stash/storage/StashObject.scala
@@ -2,7 +2,7 @@ package org.stash.storage
 
 import java.io.Serializable
 
-class StashObject(var key:String, var flags:Int, var expire:Int, var size:Int)
+class StashObject(var key:String, var flags:Int, var expire:Long, var size:Int)
     extends Serializable {
 
     var bytes: Array[Byte] = _
